+++ semo-api/src/main/java/sandbox/semo/application/common/exception/CommonErrorCode.java
