+++ semo-api/src/main/java/sandbox/semo/application/common/exception/ErrorCode.java
