+++ semo-api/src/main/java/sandbox/semo/application/member/exception/MemberBusinessException.java
