+++ semo-api/src/main/java/sandbox/semo/application/member/service/MemberService.java
@@ -1,0 +1,9 @@
+package sandbox.semo.application.member.service;
+
+import sandbox.semo.domain.member.dto.request.MemberRegister;
+
+public interface MemberService {
+
+    void register(MemberRegister request);
+
+}
