+++ semo-api/src/main/java/sandbox/semo/application/security/authentication/MemberPrincipalDetails.java
@@ -1,0 +1,33 @@
+package sandbox.semo.application.security.authentication;
+
+import java.util.Collection;
+import java.util.Collections;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import sandbox.semo.domain.member.entity.Member;
+
+@Getter
+@RequiredArgsConstructor
+public class MemberPrincipalDetails implements UserDetails {
+
+    private final Member member;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return Collections.singletonList(new SimpleGrantedAuthority(member.getRole().name()));
+    }
+
+    @Override
+    public String getPassword() {
+        return member.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return member.getLoginId();
+    }
+
+}
