+++ semo-api/src/test/java/sandbox/semo/application/SemoServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package sandbox.semo.application;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SemoServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
