+++ semo-batch/src/main/java/sandbox/semo/batch/SemoBatchApplication.java
@@ -1,0 +1,13 @@
+package sandbox.semo.batch;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SemoBatchApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SemoBatchApplication.class, args);
+    }
+
+}
