+++ semo-core/src/main/java/sandbox/semo/domain/company/repository/CompanyRepository.java
@@ -1,0 +1,8 @@
+package sandbox.semo.domain.company.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import sandbox.semo.domain.company.entity.Company;
+
+public interface CompanyRepository extends JpaRepository<Company, Long> {
+
+}
