+++ semo-core/src/main/java/sandbox/semo/domain/member/entity/Member.java
@@ -1,0 +1,61 @@
+package sandbox.semo.domain.member.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import sandbox.semo.domain.common.entity.BaseTime;
+import sandbox.semo.domain.company.entity.Company;
+
+
+@Entity
+@Getter
+@Table(name = "MEMBERS")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Member extends BaseTime {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "MEMBER_SEQ_GEN")
+    @SequenceGenerator(name = "MEMBER_SEQ_GEN", sequenceName = "MEMBER_SEQ", allocationSize = 1)
+    @Column(name = "MEMBER_ID", nullable = false)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "COMPANY_ID", nullable = false)
+    private Company company;
+
+    @Column(name = "OWNER_NAME", nullable = false, length = 30)
+    private String ownerName;
+
+    @Column(name = "LOGIN_ID", nullable = false, unique = true, length = 50)
+    private String loginId;
+
+    @Column(name = "PASSWORD", nullable = false)
+    private String password;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "ROLE", nullable = false, length = 20)
+    private Role role;
+
+    @Builder
+    public Member(Company company, String ownerName, String loginId, String password, Role role) {
+        this.company = company;
+        this.ownerName = ownerName;
+        this.loginId = loginId;
+        this.password = password;
+        this.role = role;
+    }
+
+}
