+++ semo-core/src/main/java/sandbox/semo/domain/member/entity/Role.java
@@ -1,0 +1,5 @@
+package sandbox.semo.domain.member.entity;
+
+public enum Role {
+    SUPER, ADMIN, USER
+}
