+++ semo-core/src/main/java/sandbox/semo/domain/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package sandbox.semo.domain.member.repository;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import sandbox.semo.domain.member.entity.Member;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByLoginId(String loginId);
+    Optional<Member> findByLoginIdAndDeletedAtIsNull(String loginId);
+
+}
