--- src/main/java/sandbox/semo/SemoSpringbootServerApplication.java
@@ -1,15 +1,0 @@
-package sandbox.semo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class SemoSpringbootServerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SemoSpringbootServerApplication.class, args);
-    }
-
-}
