--- src/main/java/sandbox/semo/common/config/SecurityConfig.java
@@ -1,71 +1,0 @@
-package sandbox.semo.common.config;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import sandbox.semo.security.authentication.MemberAuthProvider;
-
-@Log4j2
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final MemberAuthProvider memberAuthProvider;
-    private final AuthenticationSuccessHandler successHandler;
-    private final AuthenticationFailureHandler failureHandler;
-    private final AuthenticationEntryPoint entryPoint;
-
-    @Autowired
-    public void configure(AuthenticationManagerBuilder auth) {
-        auth.authenticationProvider(memberAuthProvider);
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .csrf(AbstractHttpConfigurer::disable)
-                .cors(AbstractHttpConfigurer::disable)
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers(
-                                "/",
-                                "/api/v1/**"
-                        ).permitAll()
-                        .anyRequest().authenticated())
-                .formLogin(form -> form
-                        .loginProcessingUrl("/api/v1/login")
-                        .successHandler(successHandler)
-                        .failureHandler(failureHandler)
-                        .permitAll())
-                .logout(exit -> exit
-                        .logoutUrl("/api/v1/logout")
-                        .deleteCookies("JSESSIONID")
-                        .invalidateHttpSession(true)
-                        .logoutSuccessHandler((
-                                (request, response, authentication)
-                                        -> SecurityContextHolder.clearContext()
-                        )))
-                .exceptionHandling(ex -> ex
-                        .authenticationEntryPoint(entryPoint));
-        return http.build();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-}
