--- src/main/java/sandbox/semo/company/repository/CompanyRepository.java
@@ -1,8 +1,0 @@
-package sandbox.semo.company.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import sandbox.semo.company.entity.Company;
-
-public interface CompanyRepository extends JpaRepository<Company, Long> {
-
-}
