--- src/main/java/sandbox/semo/member/entity/Role.java
@@ -1,5 +1,0 @@
-package sandbox.semo.member.entity;
-
-public enum Role {
-    SUPER, ADMIN, USER
-}
