--- src/main/java/sandbox/semo/member/service/MemberService.java
@@ -1,9 +1,0 @@
-package sandbox.semo.member.service;
-
-import sandbox.semo.member.dto.request.MemberRegister;
-
-public interface MemberService {
-
-    void register(MemberRegister request);
-
-}
