--- src/main/java/sandbox/semo/security/authentication/MemberPrincipalDetails.java
@@ -1,33 +1,0 @@
-package sandbox.semo.security.authentication;
-
-import java.util.Collection;
-import java.util.Collections;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import sandbox.semo.member.entity.Member;
-
-@Getter
-@RequiredArgsConstructor
-public class MemberPrincipalDetails implements UserDetails {
-
-    private final Member member;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Collections.singletonList(new SimpleGrantedAuthority(member.getRole().name()));
-    }
-
-    @Override
-    public String getPassword() {
-        return member.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return member.getLoginId();
-    }
-
-}
