--- src/test/java/sandbox/semo/SemoSpringbootServerApplicationTests.java
@@ -1,13 +1,0 @@
-package sandbox.semo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SemoSpringbootServerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
